services/consumption/telegram-service/requirements.txt
# telegram-service 依赖
python-telegram-bot>=20.0
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
